Add a (broken) playbook which adds a file to test against.
These simple tasks represent the configuration that you would be
doing in your normal playbook. In this case, we’re just touching
a file which we can then test against later.

You'll note that on line 6, this playbook actually sets the file as
"absent." This is because whenever we're writing a test, we want to
make sure it works by making sure it is Red, or broken. For more
on this process, google for "Red Green Refactor" and you'll find many
resources on the philosophy behind this.

See the docker image https://github.com/William-Yeh/docker-ansible
which explains that the base docker image we're using here defaults
to a playbook location of `./playbook.yml`. You can change this
following his instructions.